requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
astroid==1.6.6            # via pylint
coverage==5.3             # via -r requirements/test.in
ddt==1.4.1                # via -r requirements/test.in
django-threadlocals==0.10  # via -r requirements/base.txt
isort==4.3.21             # via -c requirements/constraints.txt, pylint
lazy-object-proxy==1.5.1  # via astroid
mccabe==0.6.1             # via pylint
mock==3.0.5               # via -c requirements/constraints.txt, -r requirements/test.in
pycodestyle==2.5.0        # via -c requirements/constraints.txt, -r requirements/test.in
pylint==1.9.3             # via -c requirements/constraints.txt, -r requirements/test.in
pytz==2020.1              # via -r requirements/base.txt, django
six==1.15.0               # via astroid, mock, pylint
sqlparse==0.3.1           # via -c requirements/constraints.txt, -r requirements/base.txt, django
testfixtures==6.15.0      # via -r requirements/test.in
wrapt==1.12.1             # via astroid
